refactor: 금액 계산시 BigDecimal 사용

+++ src/main/kotlin/lotto/domain/LottoStatistics.kt
@@ -1,13 +1,15 @@
 package lotto.domain
 
+import java.math.BigDecimal
+
 class LottoStatistics(private val money: Money, results: Map<Winning, Int>) {
-    var totalAmount: Double = 0.0
+    var totalAmount: BigDecimal = 0.toBigDecimal()
 
     init {
         results.forEach { (winning, winningCount) ->
-            totalAmount += winning.winningAmount * winningCount
+            totalAmount += (winning.winningAmount * winningCount).toBigDecimal()
         }
     }
 
-    fun getYield(): Double = totalAmount / money.price
+    fun getYield(): BigDecimal = totalAmount / money.price.toBigDecimal()
 }
